+++ .github/workflows/sonar.yml
@@ -27,6 +27,7 @@ jobs:
           # Ensure we are building the branch and not the branch after being merged on develop
           # https://github.com/actions/checkout/issues/881
           ref: ${{ github.event_name == 'pull_request' && github.event.pull_request.head.sha || github.ref }}
+          lfs: true
       - name: Use JDK 21
         uses: actions/setup-java@v4
         with:
