+++ tests/konsist/src/test/kotlin/io/element/android/tests/konsist/KonsistPreviewTest.kt
@@ -72,6 +72,7 @@ class KonsistPreviewTest {
                 "IconsCompoundPreview",
                 "IconsOtherPreview",
                 "MarkdownTextComposerEditPreview",
+                "MatrixBadgeAtomExternalPreview",
                 "MatrixBadgeAtomPositivePreview",
                 "MatrixBadgeAtomNeutralPreview",
                 "MatrixBadgeAtomNegativePreview",
@@ -103,6 +104,7 @@ class KonsistPreviewTest {
                 "RoomMemberListViewBannedPreview",
                 "SasEmojisPreview",
                 "SecureBackupSetupViewChangePreview",
+                "SelectedUserExternalPreview",
                 "SelectedUserCannotRemovePreview",
                 "TextComposerAddCaptionPreview",
                 "TextComposerCaptionPreview",
